// admin_frontend/src/lib/api/models/ProductCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProductCreate = {
    title: string;
    slug: string;
    tea_type?: (string | null);
    description?: (string | null);
    lore_description?: (string | null);
    brewing_guide?: (Record<string, any> | null);
    category_id: number;
    is_active?: (boolean | null);
    seo_title?: (string | null);
    seo_description?: (string | null);
    seo_keywords?: (string | null);
};