// admin_frontend/src/lib/api/models/Category.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Category = {
    name: string;
    slug: string;
    description?: (string | null);
    image?: (string | null);
    parent_id?: (number | null);
    seo_title?: (string | null);
    seo_description?: (string | null);
    is_active?: (boolean | null);
    id: number;
};